+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -8,10 +8,10 @@ import java.util.Collection;
 public interface UserStorage {
     void add(User user) throws IncorrectIdException;
     void update(User user) throws IncorrectIdException;
-    User find(int id) throws IncorrectIdException;
+    User find(long id) throws IncorrectIdException;
     Collection<User> getAll();
-    void addFriend(int userId, int friendId) throws IncorrectIdException;
-    void removeFriend(int userId, int friendId) throws IncorrectIdException;
-    Collection<User> getCommonFriends(int user1, int user2) throws IncorrectIdException;
-    boolean contains(int id);
+    void addFriend(long userId, long friendId) throws IncorrectIdException;
+    void removeFriend(long userId, long friendId) throws IncorrectIdException;
+    Collection<User> getCommonFriends(long user1, long user2) throws IncorrectIdException;
+    boolean contains(long id);
 }
